Move mock statement data from model to mock-data module.
